+++ ui/src/components/AdminOrderModal.jsx
@@ -1,17 +1,17 @@
 import React, { useState } from "react";
-import ParcelSizeSummaryCard from "./SummaryCards/ParcelSizeSummaryCard";
-import RecipientSummaryCard from "./SummaryCards/RecipientSummaryCard";
-import SenderSummaryCard from "./SummaryCards/SenderSummaryCard";
-import { Dialog, Select, MenuItem, Button } from "@material-ui/core";
+import { Dialog, Select, MenuItem } from "@material-ui/core";
 import "../styles/OrderModal.css";
-import StatusIcon from "./StatusIcon";
 import { ADMIN_ORDERS } from "../api/config";
 import axiosInstance from "../api/axiosInstance";
 import useMessage from "../hooks/messages";
+import OrderModalSummary from "./OrderModalSummary";
+import { getOrderStatusList } from "../services/orderStatus";
+import LoadingButton from "./LoadingButton";
 
-export default function OrderModal(props) {
+export default function AdminOrderModal(props) {
   const [status, setStatus] = useState(props.order.status);
   const { displayError, displaySuccess } = useMessage();
+  const [isChangingStatus, setIsChangingStatus] = useState(false);
 
   const handleChange = (event) => {
     setStatus(event.target.value);
@@ -19,6 +19,7 @@ export default function OrderModal(props) {
 
   const handleSave = () => {
     (async function fetchData() {
+      setIsChangingStatus(true);
       try {
         await axiosInstance.put(
           ADMIN_ORDERS + "/" + props.order.id + "/status",
@@ -28,6 +29,8 @@ export default function OrderModal(props) {
         props.handleStatusChange(props.order.id, status);
       } catch (e) {
         displayError("Failed to change status");
+      } finally {
+        setIsChangingStatus(false);
       }
     })();
   };
@@ -35,22 +38,7 @@ export default function OrderModal(props) {
   return (
     <Dialog open={props.open} onClose={() => props.onClose(props.order)}>
       <div className="modal-wrapper">
-        <div className="order-status-wrapper order-info-field-wrapper">
-          <span className="order-status-title ">Order Status:</span>
-          <StatusIcon status={props.order.status} fontSize={30} />
-          <span className="description-text">{props.order.status}</span>
-        </div>
-        <SenderSummaryCard
-          sender={props.order.senderInfo}
-          pickUpDate={props.order.pickupDateTime}
-        />
-        <RecipientSummaryCard recipient={props.order.recipientInfo} />
-        <ParcelSizeSummaryCard
-          selectedPackageSize={props.order.packageOption.packageSize}
-          selectedPackageType={props.order.packageOption.packageType.title}
-          attributes={props.order.attributes}
-          price={props.order.totalPrice}
-        />
+        <OrderModalSummary order={props.order} />
         <div className="admin-status">
           <span className="admin-modal-status">Status: </span>
           <Select
@@ -58,20 +46,24 @@ export default function OrderModal(props) {
             onChange={handleChange}
             className="admin-modal-select"
           >
-            <MenuItem value={"NEW"}>NEW</MenuItem>
-            <MenuItem value={"DELIVERED"}>DELIVERED</MenuItem>
-            <MenuItem value={"CANCELED"}>CANCELED</MenuItem>
+            {getOrderStatusList().map((status, index) => (
+              <MenuItem value={status} key={index}>
+                {status}
+              </MenuItem>
+            ))}
           </Select>
         </div>
         <div className="admin-modal-buttons">
-          <Button
+          <LoadingButton
             color="primary"
             variant="contained"
             className="form-button admin-modal-save"
             onClick={handleSave}
+            loading={isChangingStatus}
+            disabled={isChangingStatus}
           >
             Save
-          </Button>
+          </LoadingButton>
         </div>
       </div>
     </Dialog>
