+++ ui/src/services/orderStatus.js
@@ -1,0 +1,5 @@
+const STATUS = ["NEW", "DELIVERED", "CANCELED"];
+
+export function getOrderStatusList() {
+  return STATUS;
+}
